Add Oref-GPT (#237)
* Added Home Assistant Red Alert app

Added this Home Alert integration for the Red Alert / Pikud HaOref system (language: Python)

* Revert "Added Home Assistant Red Alert app"

This reverts commit dd7b5be032cdcf78e9903fab7209b35189277d05.

* Added Red Alert (Home Assistant integration)

Added the Red Alert Home Assistant integration (one of two!). Tried my best to avoid whitespaces in the TOS!

* Fixed spacing issue (hopefully!)

Used a different IDE with a different setting to remove the problematic space handling

* More layout fixes

Fixed the spacing on the addition itself and corrected the link for the stars

* Added link to HA repo

Set up a repo for Home Assistant components. Added a link to it.

* Revert "Added link to HA repo"

This reverts commit 240c3b5e5f6ba48caf0b8364dad24cfaf78f1a95.

* Add Oref-GPT to HTML section

---------

Co-authored-by: Daniel Rosehill <github@danielrosehill.co.il>